move useLocationQuery
